requirements.txt
rich
google-generativeai
pytubefix
whisper
bs4
requests
python-dotenv
